legacy/old-project-files/src/scrapers/main.py
#!/usr/bin/env python3
"""
Main scraper service for OpenPolicy Merge
Coordinates all scraping activities across different jurisdictions
"""

import os
import time
import logging
import schedule
from datetime import datetime
import redis
import psycopg2
from psycopg2.extras import RealDictCursor

# Configure logging
logging.basicConfig(
    level=logging.INFO,
    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s'
)
logger = logging.getLogger(__name__)

class ScraperService:
    def __init__(self):
        self.redis_client = redis.Redis(
            host=os.getenv('REDIS_HOST', 'redis'),
            port=int(os.getenv('REDIS_PORT', 6379)),
            decode_responses=True
        )
        
        self.db_connection = None
        self.connect_database()
    
    def connect_database(self):
        """Connect to PostgreSQL database"""
        try:
            self.db_connection = psycopg2.connect(
                host=os.getenv('DB_HOST', 'postgres'),
                port=int(os.getenv('DB_PORT', 5432)),
                database=os.getenv('DB_NAME', 'openpolicy'),
                user=os.getenv('DB_USER', 'postgres'),
                password=os.getenv('DB_PASSWORD', 'password')
            )
            logger.info("Connected to database successfully")
        except Exception as e:
            logger.error(f"Failed to connect to database: {e}")
    
    def scrape_federal_data(self):
        """Scrape federal government data"""
        try:
            logger.info("Starting federal data scraping...")
            # TODO: Implement federal scraping logic
            self.redis_client.set('scrape:federal:last_run', datetime.now().isoformat())
            logger.info("Federal data scraping completed")
        except Exception as e:
            logger.error(f"Federal scraping failed: {e}")
    
    def scrape_provincial_data(self):
        """Scrape provincial government data"""
        try:
            logger.info("Starting provincial data scraping...")
            # TODO: Implement provincial scraping logic
            self.redis_client.set('scrape:provincial:last_run', datetime.now().isoformat())
            logger.info("Provincial data scraping completed")
        except Exception as e:
            logger.error(f"Provincial scraping failed: {e}")
    
    def scrape_municipal_data(self):
        """Scrape municipal government data"""
        try:
            logger.info("Starting municipal data scraping...")
            # TODO: Implement municipal scraping logic
            self.redis_client.set('scrape:municipal:last_run', datetime.now().isoformat())
            logger.info("Municipal data scraping completed")
        except Exception as e:
            logger.error(f"Municipal scraping failed: {e}")
    
    def run_scheduled_scraping(self):
        """Run all scheduled scraping tasks"""
        logger.info("Running scheduled scraping tasks...")
        
        # Schedule daily scraping at 2 AM
        schedule.every().day.at("02:00").do(self.scrape_federal_data)
        schedule.every().day.at("03:00").do(self.scrape_provincial_data)
        schedule.every().day.at("04:00").do(self.scrape_municipal_data)
        
        while True:
            schedule.run_pending()
            time.sleep(60)
    
    def run_immediate_scraping(self):
        """Run immediate scraping for all jurisdictions"""
        logger.info("Running immediate scraping...")
        self.scrape_federal_data()
        self.scrape_provincial_data()
        self.scrape_municipal_data()
        logger.info("Immediate scraping completed")

def main():
    """Main entry point"""
    logger.info("Starting OpenPolicy Merge Scraper Service...")
    
    scraper = ScraperService()
    
    # Run immediate scraping first
    scraper.run_immediate_scraping()
    
    # Then start scheduled scraping
    scraper.run_scheduled_scraping()

if __name__ == "__main__":
    main()
