legacy/old-project-files/src/scrapers/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
selenium==4.15.2
psycopg2-binary==2.9.7
redis==5.0.1
celery==5.3.4
scrapy==2.11.0
pandas==2.1.3
numpy==1.25.2
python-decouple==3.8
schedule==1.2.0
